--- src/components/Camera.css
@@ -1,22 +1,0 @@
-.camera-section {
-    position: relative;
-    width: 50%;
-    border-right: 1px solid rgba(139, 69, 19, 0.1);
-    background-color: #f8f5f1;
-  }
-  
-  .camera-video {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-  
-  .camera-error {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-    background-color: #f5f1ec;
-    color: red;
-    font-weight: bold;
-  }
