--- src/components/ChatContainer.css
@@ -1,112 +1,0 @@
-.chat-container {
-    display: flex;
-    flex-direction: column;
-    height: 100%;
-  }
-  
-  .chat-header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 16px;
-  }
-  
-  .chat-header-logo {
-    height: 100px;
-    width: auto;
-  }
-  
-  .chat-button {
-    border-radius: 20px;
-    background-color: #8B4513;
-    padding: 8px 24px;
-    color: white;
-    font-size: 1.2em;
-    border: none;
-    cursor: pointer;
-    transition: all 0.3s ease;
-    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
-  }
-  
-  .chat-button:hover {
-    background-color: #5D3311;
-    transform: translateY(-2px);
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
-  }
-  
-  .chat-messages {
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    overflow-y: auto;
-    background-color: var(--chat-bg-color);
-    padding: 16px;
-    margin-bottom: 16px;
-    border-radius: 8px;
-  }
-  
-  .chat-message {
-    max-width: 70%;
-    padding: 12px;
-    margin-bottom: 10px;
-    border-radius: 18px;
-    word-wrap: break-word;
-    animation: fadeIn 0.3s ease-in-out;
-  }
-  
-  .user-message {
-    align-self: flex-end;
-    background-color: #e3f2fd;
-    color: #000;
-    margin-left: auto;
-    border-bottom-right-radius: 4px;
-  }
-  
-  .bot-message {
-    align-self: flex-start;
-    background-color: #f0f0f0;
-    color: #000;
-    margin-right: auto;
-    border-bottom-left-radius: 4px;
-  }
-  
-  .input-container {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-  }
-  
-  .input-field {
-    flex-grow: 1;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 4px;
-    font-size: 1.2em;
-    transition: border-color 0.3s;
-  }
-  
-  .input-field:focus {
-    outline: none;
-    border-color: #8B4513;
-  }
-  
-  .send-button {
-    padding: 8px 20px;
-    background-color: #8B4513;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    transition: background-color 0.3s, transform 0.3s;
-    font-size: 1.2em;
-  }
-  
-  .send-button:hover {
-    background-color: #5D3311;
-    transform: translateY(-2px);
-  }
-  
-  @keyframes fadeIn {
-    from { opacity: 0; transform: translateY(10px); }
-    to { opacity: 1; transform: translateY(0); }
-  }
