+++ src/components/Header.css
@@ -1,0 +1,17 @@
+#recordButton {
+    font-size: 1em;
+    padding: 8px 16px;
+    border: none;
+    border-radius: 20px;
+    background-color: #1a73e8;
+    color: white;
+    cursor: pointer;
+    transition: all 0.3s ease;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+}
+
+#recordButton:hover {
+    background-color: #1558d6;
+    transform: translateY(-2px);
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
+}
