+++ src/components/Waveform.js
@@ -1,0 +1,87 @@
+// Waveform.js
+import React, { useEffect, useRef } from 'react';
+
+const Waveform = ({ analyser, isActive }) => {
+  const canvasRef = useRef(null);
+  const animationFrameRef = useRef(null);
+
+  useEffect(() => {
+    const canvas = canvasRef.current;
+    const canvasCtx = canvas.getContext('2d');
+
+    // Make canvas responsive
+    const resizeCanvas = () => {
+      const container = canvas.parentElement;
+      canvas.width = container.clientWidth;
+      canvas.height = 100;
+    };
+
+    resizeCanvas();
+    window.addEventListener('resize', resizeCanvas);
+
+    const bufferLength = analyser.frequencyBinCount;
+    const dataArray = new Uint8Array(bufferLength);
+
+    const drawWaveform = () => {
+      if (!isActive) return;
+
+      analyser.getByteTimeDomainData(dataArray);
+
+      const width = canvas.width;
+      const height = canvas.height;
+
+      canvasCtx.fillStyle = 'rgb(255, 255, 255)';
+      canvasCtx.fillRect(0, 0, width, height);
+
+      // 그라데이션 생성
+      const gradient = canvasCtx.createLinearGradient(0, 0, width, 0);
+      gradient.addColorStop(0, '#FF0000');    // 시작 색상
+      gradient.addColorStop(0.5, '#00FF00');  // 중간 색상
+      gradient.addColorStop(1, '#0000FF');    // 끝 색상
+
+      canvasCtx.lineWidth = 2;
+      canvasCtx.strokeStyle = gradient;  // 그라데이션 적용
+      canvasCtx.beginPath();
+
+      const sliceWidth = width * 1.0 / bufferLength;
+      let x = 0;
+
+      for (let i = 0; i < bufferLength; i++) {
+        const v = dataArray[i] / 128.0;
+        const y = v * height / 2;
+
+        if (i === 0) {
+          canvasCtx.moveTo(x, y);
+        } else {
+          canvasCtx.lineTo(x, y);
+        }
+
+        x += sliceWidth;
+      }
+
+      canvasCtx.lineTo(canvas.width, canvas.height / 2);
+      canvasCtx.stroke();
+
+      animationFrameRef.current = requestAnimationFrame(drawWaveform);
+    };
+
+    if (isActive) {
+      drawWaveform();
+    }
+
+    return () => {
+      window.removeEventListener('resize', resizeCanvas);
+      if (animationFrameRef.current) {
+        cancelAnimationFrame(animationFrameRef.current);
+      }
+    };
+  }, [analyser, isActive]);
+
+  return (
+    <div style={{ width: '100%', height: '100px' }}>
+      <canvas ref={canvasRef} style={{ width: '100%', height: '100%' }} />
+    </div>
+  );
+};
+
+export default Waveform;
