--- src/components/common/Header.css
@@ -1,14 +1,0 @@
-/* components/common/Header.css */
-.header {
-    display: flex;
-    height: 131px;
-    align-items: center;
-    background-color: var(--secondary-color);
-    padding: 0 22px;
-    z-index: 2;
-}
-
-.header-logo {
-    height: 95px;
-    width: auto;
-}
