--- src/styles/HomePage.css
@@ -1,209 +1,0 @@
-@import url('https://cdn.rawgit.com/kattergil/NotoSerifKR-Web/76eb9ebf/stylesheet/NotoSerif-Web.css');
-
-/* HomePage.css */
-:root {
-    --primary-color: #DFD9CE;
-    --secondary-color: #F4F1EA;
-    --hover-color: #724023;
-}
-
-.home-container {
-    min-height: 100vh;
-    background-color: var(--primary-color);
-    display: flex;
-    flex-direction: column;
-    overflow: hidden;
-}
-
-.home-main {
-    position: relative;
-    height: calc(100vh - 100px);
-    width: 100%;
-    padding: 2vw 4vw;
-}
-
-/* Title Section */
-.title-section {
-    position: absolute;
-    left: 4vw;
-    top: calc(10px);
-    z-index: 2;
-    max-width: 50%;
-}
-
-/* HomePage.css */
-.main-title {
-    font-family: 'Noto Serif KR', serif;
-    font-size: clamp(48px, 3vw, 48px);
-    font-weight: 1000;
-    color: #000;
-    margin-bottom: 1vh;
-    line-height: 1.3;
-}
-
-
-
-.sub-title {
-    font-family: 'Noto Serif KR', serif;
-    font-size: clamp(64px, 4vw, 64px);
-    font-weight: 1500;
-    color: #000;
-    line-height: 1.2;
-}
-
-/* Books Section */
-.books-section {
-    position: relative;
-    height: 100%;
-    width: 100%;
-}
-
-/* Left Book Container */
-.left-book-container {
-    position: absolute;
-    left: -5vw;
-    bottom: 2vh;
-    z-index: 1;
-}
-
-/* Right Book Container */
-.right-book-container {
-    position: absolute;
-    right: -2vw;
-    top: -20vh;
-    /* top 위치를 5vh로 줄여서 header 바로 아래로 이동 */
-    z-index: 1;
-    padding-right: 6%;
-    /* z-index를 조정하여 다른 요소 위로 표시 */
-}
-
-/* Books */
-/* 오른쪽 책 이미지 크기 조정 */
-/* 오른쪽 책과 왼쪽 책 모두 동일한 크기 강제 설정 */
-/* 왼쪽 책 */
-.left-book {
-    width: 1.843*40vw; /* 양쪽 책의 너비를 동일하게 고정 */
-    height: 40vw; /* 높이는 비율에 따라 자동 조정 */
-    min-width: calc(1093px);
-    min-height: calc(593px*1.05);
-    max-width: calc(1093px * 1.2); /* 최대 너비 1093px의 1.3배 */
-    max-height: calc(593px * 1.2); /* 최대 높이 593px의 1.3배 */
-    filter: drop-shadow(0 2vh 4vh rgba(0, 0, 0, 0.1));
-    transition: transform 0.3s ease;
-}
-
-/* 오른쪽 책 */
-.right-book {
-    width: 50vw; /* 양쪽 책의 너비를 동일하게 고정 */
-    height: 50vw*1.072; /* 높이는 비율에 따라 자동 조정 */
-    min-width: calc(789px);
-    min-height: calc(846px);
-    max-width : calc(789px*1.2);
-    max-height: calc(846px*1.2);
-    filter: drop-shadow(0 2vh 4vh rgba(0, 0, 0, 0.1));
-    transition: transform 0.3s ease;
-}
-
-
-  
-/* Book Overlay */
-/* Book Overlay */
-.book-overlay-left {
-    position: absolute;
-    top: 76%; /* Adjust as needed to position '읽으러 가기' over the book */
-    left: 39.5%;
-    transform: translate(-50%, -50%);
-    color: #fff;
-    text-align: center;
-    width: auto;
-    padding: 0 20px;
-}
-
-.book-overlay-right {
-    position: absolute;
-    top: 35%; /* Adjust as needed to position '읽으러 가기' over the book */
-    right: 20.5%;
-    transform: translate(-50%, -50%);
-    color: #fff;
-    text-align: center;
-    width: auto;
-    padding: 0 20px;
-}
-
-
-.book-title {
-    font-size: clamp(20px, 2.5vw, 36px);
-    font-weight: 500;
-    margin-bottom: 0vh;
-    color: #FFFFFF;
-    text-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.book-link {
-    display: inline-flex;
-    align-items: center;
-    justify-content: center;
-    gap: 1.5vw;
-    background-color: rgba(250, 247, 242, 0.9);
-    color: #8B7E74;
-    padding: 2.5vh 3vw;
-    border-radius: 100px;
-    text-decoration: none;
-    font-weight: 500;
-    font-size: clamp(14px, 1vw, 18px);
-    transition: all 0.3s ease;
-    backdrop-filter: blur(4px);
-}
-
-.book-link:hover {
-    background-color: rgba(240, 235, 227, 0.95);
-    transform: translateY(-2px);
-}
-
-/* Responsive Design */
-@media (max-aspect-ratio: 1/1) {
-    .home-main {
-        padding: 4vw;
-        height: auto;
-        min-height: calc(100vh - 80px);
-    }
-
-    .title-section {
-        position: relative;
-        left: 0;
-        top: 0;
-        margin-bottom: 4vh;
-        max-width: 100%;
-    }
-
-    .books-section {
-        display: flex;
-        flex-direction: column;
-        gap: 4vh;
-    }
-/* 
-    .left-book-container,
-    .right-book-container {
-        position: relative;
-        left: 0;
-        right: 0;
-        top: 0;
-        margin: 0 auto;
-    } */
-
-    /* .left-book,
-    .right-book {
-        width: 100vw;
-        max-width: 700px;
-        margin: 0 auto;
-        display: block;
-    } */
-}
-
-/* 매우 큰 화면에서의 최대 크기 제한 */
-@media (min-width: 2560px) {
-    .home-main {
-        max-width: 2560px;
-        margin: 0 auto;
-    }
-}
