--- src/styles/PageFlip.css
@@ -1,68 +1,0 @@
-/* PageFlip.css */
-.app-container {
-    position: relative;
-    width: 100vw;
-    height: 100vh;
-    overflow: hidden;
-    perspective: 1500px;
-    background: #f5f5f5;
-  }
-  
-  .page-container {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    transform-origin: left center;
-    transform-style: preserve-3d;
-    background: white;
-    box-shadow: 0 0 20px rgba(0,0,0,0.1);
-  }
-  
-  .page-content {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    backface-visibility: hidden;
-  }
-  
-  .page-fold {
-    position: absolute;
-    top: 0;
-    right: 0;
-    width: 50%;
-    height: 100%;
-    background: linear-gradient(to right, 
-      rgba(0,0,0,0) 0%,
-      rgba(0,0,0,0.1) 50%,
-      rgba(0,0,0,0.15) 100%);
-    transform-origin: left;
-    opacity: 0;
-    transition: opacity 0.4s;
-  }
-  
-  .page-container:hover .page-fold {
-    opacity: 1;
-  }
-  
-  /* 페이지가 넘어갈 때 그림자 효과 */
-  .page-container::after {
-    content: '';
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    background: linear-gradient(to right,
-      rgba(0,0,0,0.2),
-      rgba(0,0,0,0) 20%,
-      rgba(0,0,0,0) 80%,
-      rgba(0,0,0,0.2)
-    );
-    opacity: 0;
-    transition: opacity 0.4s;
-    pointer-events: none;
-  }
-  
-  .page-container:hover::after {
-    opacity: 1;
-  }
